requirements.txt
gTTS==1.2.2
requests==2.18.4
tinydb==3.5.0
python-telegram-bot==8.0
